# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an api that make authorization",
        "title": "Authorization",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/session": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Refresh a user session and set in the response header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Refresh session",
                "responses": {
                    "200": {
                        "description": "user session refreshed successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "401": {
                        "description": "user session has expired",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a user session and set in the response header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Create session",
                "parameters": [
                    {
                        "description": "user params",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserSessionPartial"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "session created successfully",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "400": {
                        "description": "an invalid user param was sent",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "404": {
                        "description": "user does not exist",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/server.sent"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.UserSessionPartial": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "server.sent": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "apiKey",
            "name": "Session",
            "in": "header"
        }
    }
}